Move retrotype_cli into retrotype directory.
